Module Paf_mirage

module type S = sig ... end

module Make (Stack : Tcpip.Tcp.S) : S with type stack = Stack.t and type ipaddr = Stack.ipaddr

Client implementation.

The client implementation of Paf_mirage does not strictly need a functor. Indeed, the client was made in the sense of mimic. The user should provide a Mimic.ctx which generate a paf_transmission. By this way, the run function is able to introspect the used protocol (regardless its implementation) and do the ALPN challenge with the server.

type transmission = [

| `Clear

| `TLS of string option

]

val paf_transmission : transmission Mimic.value

val run : ctx:Mimic.ctx -> (Ipaddr.t * int) option Alpn.client_handler -> [ `V1 of Httpaf.Request.t | `V2 of H2.Request.t ] -> (Alpn.alpn_response, [> Mimic.error ]) Stdlib.result Lwt.t

run ~ctx handler req sends an HTTP request (H2 or HTTP/1.1) to a peer which can be reached via the given Mimic's ctx. If the connection is recognized as a tls_protocol, we proceed an ALPN challenge between what the user chosen and what the peer can handle. Otherwise, we send a simple HTTP/1.1 request or a h2c request.